+++ Source/DafnyLanguageServer/Workspace/ITextDocumentLoader.cs
@@ -32,7 +32,6 @@ namespace Microsoft.Dafny.LanguageServer.Workspace {
     /// <exception cref="System.OperationCanceledException">Thrown when the cancellation was requested before completion.</exception>
     /// <exception cref="System.ObjectDisposedException">Thrown if the cancellation token was disposed before the completion.</exception>
     Task<DafnyDocument> LoadAsync(DocumentTextBuffer textDocument, CancellationToken cancellationToken);
-    Task<DafnyDocument> PrepareVerificationTasksAsync(DafnyDocument loaded, CancellationToken cancellationToken);
 
     /// <summary>
     /// Publishes the verification diagnostics (VerificationTree) for the given document
